+++ packages/cap-config-writer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/cap-config-writer
 
+## 0.10.11
+
+### Patch Changes
+
+-   Updated dependencies [4fef16a]
+    -   @sap-ux/fiori-generator-shared@0.12.10
+
+## 0.10.10
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.3
+-   @sap-ux/fiori-generator-shared@0.12.9
+
 ## 0.10.9
 
 ### Patch Changes