+++ packages/cf-deploy-config-inquirer/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/cf-deploy-config-inquirer
 
+## 0.3.17
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.15
+
+## 0.3.16
+
+### Patch Changes
+
+-   @sap-ux/inquirer-common@0.7.14
+
+## 0.3.15
+
+### Patch Changes
+
+-   ff6ff97: Update the MTA ID error message.
+
 ## 0.3.14
 
 ### Patch Changes