+++ packages/fiori-annotation-api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sap-ux/fiori-annotation-api
 
+## 0.6.4
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.3
+-   @sap-ux/cds-odata-annotation-converter@0.6.3
+
 ## 0.6.3
 
 ### Patch Changes