+++ packages/generator-adp/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @sap-ux/generator-adp
 
+## 0.3.30
+
+### Patch Changes
+
+-   Updated dependencies [4fef16a]
+    -   @sap-ux/fiori-generator-shared@0.12.10
+    -   @sap-ux/adp-flp-config-sub-generator@0.1.30
+    -   @sap-ux/inquirer-common@0.7.15
+    -   @sap-ux/adp-tooling@0.14.30
+
+## 0.3.29
+
+### Patch Changes
+
+-   Updated dependencies [f9ea9e3]
+    -   @sap-ux/adp-flp-config-sub-generator@0.1.29
+    -   @sap-ux/adp-tooling@0.14.29
+    -   @sap-ux/axios-extension@1.22.3
+    -   @sap-ux/system-access@0.6.8
+    -   @sap-ux/project-access@1.30.3
+    -   @sap-ux/fiori-generator-shared@0.12.9
+    -   @sap-ux/project-input-validator@0.6.4
+    -   @sap-ux/inquirer-common@0.7.14
+
+## 0.3.28
+
+### Patch Changes
+
+-   Updated dependencies [14214a3]
+    -   @sap-ux/axios-extension@1.22.2
+    -   @sap-ux/adp-flp-config-sub-generator@0.1.28
+    -   @sap-ux/adp-tooling@0.14.28
+    -   @sap-ux/system-access@0.6.7
+
 ## 0.3.27
 
 ### Patch Changes