+++ packages/project-input-validator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/project-input-validator
 
+## 0.6.4
+
+### Patch Changes
+
+-   @sap-ux/project-access@1.30.3
+
 ## 0.6.3
 
 ### Patch Changes